pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "workspace-templates"
version = "0.1.0"
description = "KDE Plasma 6 Wayland workspace template manager"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
  { name = "Codex" }
]
dependencies = [
  "PyQt6>=6.6",
  "dbus-python>=1.3.2",
  "PyGObject>=3.46"
]

[project.scripts]
workspace-templates = "workspace_templates.main:main"

[tool.setuptools]
include-package-data = true
packages = ["workspace_templates"]

[tool.setuptools.package-data]
workspace_templates = [
  "assets/*.desktop",
  "assets/kwin-shortcut/metadata.json",
  "assets/kwin-shortcut/contents/code/main.js"
]
